// solutions_backend/src/database/Interfaces/Clients.Interface.ts
export interface IcreateInterface {
    application_id: number;
    completeName: string;
    companyName: string;
    profile?: string;
    username: string;
    document: string;
    birthday?: string;
    email: string;
    telephone: string;
    password: string;
    active?: string;
    payment?: number;
    day_payment?: string;
    date_payment?: number;
    type?: number;
}